requirements.txt
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.74
PyQt6==6.5.2
tensorflow-cpu==2.13.0
pytest==7.4.0
